[finance_report_vision] fix(llm): api_key_version must be BigInteger (fingerprint overflows int32)
The key fingerprint (sha256[:4]) is an unsigned 32-bit value up to ~4.3e9, which
overflows a signed int32 Postgres column — asyncpg raised 'value out of int32
range' whenever a key hashed above 2^31 (CI hit it; local passed by luck on
smaller random fingerprints). Make llm_providers.api_key_version BigInteger.
Verified: migration applies, alembic check clean, the CI value inserts, and the
integration test passes 5/5.

+++ apps/backend/migrations/versions/0043_llm_provider_config.py
@@ -48,7 +48,7 @@ def upgrade() -> None:
             nullable=False,
         ),
         sa.Column("api_key_ciphertext", sa.String(), nullable=False),
-        sa.Column("api_key_version", sa.Integer(), nullable=False),
+        sa.Column("api_key_version", sa.BigInteger(), nullable=False),
         sa.Column("api_base", sa.String(length=500), nullable=True),
         sa.Column("created_at", sa.DateTime(timezone=True), nullable=False),
         sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False),
